requirements.txt
absl-py==0.10.0
aiocontextvars==0.2.2
aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-kms==2.16.0
attrs==20.2.0
cachetools==4.1.1
certifi==2021.5.30
cffi==1.15.1
chardet==3.0.4
click==7.1.2
contextvars==2.4
crcmod==1.7
cryptography==37.0.4
cycler==0.10.0
Cython==0.29.21
dataclasses==0.8
decorator==4.4.2
flake8==3.8.3
flake8-import-order==0.18.1
future==0.18.2
google-auth==1.22.1
google-auth-oauthlib==0.4.1
grpcio==1.32.0
idna==2.10
imageio==2.15.0
imageio-ffmpeg==0.4.7
immutables==0.18
importlib-metadata==2.0.0
iniconfig==1.0.1
jmespath==0.10.0
joblib==1.1.0
jsonpatch==1.32
jsonpointer==2.3
kiwisolver==1.2.0
llvmlite==0.36.0
loguru==0.6.0
Markdown==3.3
matplotlib==3.3.2
mccabe==0.6.1
motmetrics==1.2.0
networkx==2.5.1
numba==0.53.1
numpy==1.19.2
oauthlib==3.1.0
opencv-python==4.4.0.44
oss2==2.16.0
packaging==20.4
pandas==1.1.2
Pillow==8.4.0
pluggy==0.13.1
protobuf==3.13.0
py==1.9.0
py-cpuinfo==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycodestyle==2.6.0
pycparser==2.21
pycryptodome==3.15.0
pyflakes==2.2.0
pyparsing==2.4.7
pytest==6.1.0
pytest-benchmark==3.2.3
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==23.1.0
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.2
seaborn==0.11.2
Shapely==1.7.1
sheen==0.1.2
six==1.15.0
sklearn==0.0
tabulate==0.8.7
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorboardX==2.1
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2020.9.3
toml==0.10.1
torch==1.10.1
torchfile==0.1.0
torchvision==0.11.2
tornado==6.1
tqdm==4.49.0
tslearn==0.5.2
typing_extensions==4.1.1
urllib3==1.25.10
visdom==0.1.8.9
websocket-client==1.3.1
Werkzeug==1.0.1
xmltodict==0.12.0
yacs==0.1.8
zipp==3.2.0
